% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myadd.r
\name{myadd}
\alias{myadd}
\title{An Integer Addition Function}
\usage{
myadd(a, b = 0)
}
\arguments{
\item{a}{the first number}

\item{b}{the second number, default to 0}
}
\description{
This function allows you to add two numbers.
}
\examples{
myadd(2,3)
}
\keyword{add}
\keyword{myadd}

